+++ ComponentImpl/src/main/java/com/xiaojinzi/component/application/IComponentHostApplication.java
@@ -5,5 +5,5 @@ import com.xiaojinzi.component.support.IHost;
 /**
  * 每一个实现类都必须返回对应的 Host
  */
-public interface IComponentHostApplication extends IComponentApplication, IHost {
+public interface IComponentHostApplication extends IApplicationLifecycle, IHost {
 }