+++ ComponentImpl/src/main/java/com/xiaojinzi/component/fragment/IComponentHostFragment.java
@@ -3,7 +3,7 @@ package com.xiaojinzi.component.fragment;
 import android.support.annotation.MainThread;
 import android.support.annotation.NonNull;
 
-import com.xiaojinzi.component.support.IModuleLifecycle;
+import com.xiaojinzi.component.support.IBaseLifecycle;
 import com.xiaojinzi.component.support.IHost;
 
 import java.util.Set;
@@ -11,7 +11,7 @@ import java.util.Set;
 /**
  * 每个模块的接口,需要有生命周期的通知
  */
-public interface IComponentHostFragment extends IModuleLifecycle, IHost {
+public interface IComponentHostFragment extends IBaseLifecycle, IHost {
 
     /**
      * 主要用于检查模块和模块之间的重复