+++ ComponentImpl/src/main/java/com/xiaojinzi/component/interceptor/IComponentHostInterceptor.java
@@ -5,7 +5,7 @@ import android.support.annotation.NonNull;
 
 import com.xiaojinzi.component.impl.RouterInterceptor;
 import com.xiaojinzi.component.impl.interceptor.InterceptorBean;
-import com.xiaojinzi.component.support.IModuleLifecycle;
+import com.xiaojinzi.component.support.IBaseLifecycle;
 import com.xiaojinzi.component.support.IHost;
 
 import java.util.List;
@@ -18,7 +18,7 @@ import java.util.Set;
  * @author : xiaojinzi
  */
 public interface IComponentHostInterceptor
-        extends IComponentInterceptor, IModuleLifecycle, IHost {
+        extends IComponentInterceptor, IBaseLifecycle, IHost {
 
     /**
      * 获取全局的拦截器列表,外面使用的地方做了缓存