+++ ComponentImpl/src/main/java/com/xiaojinzi/component/service/IComponentHostService.java
@@ -1,10 +1,10 @@
 package com.xiaojinzi.component.service;
 
-import com.xiaojinzi.component.support.IModuleLifecycle;
+import com.xiaojinzi.component.support.IBaseLifecycle;
 import com.xiaojinzi.component.support.IHost;
 
 /**
  * 每个模块的接口,需要有生命周期的通知
  */
-public interface IComponentHostService extends IModuleLifecycle, IHost {
+public interface IComponentHostService extends IBaseLifecycle, IHost {
 }