+++ ComponentImpl/src/main/java/com/xiaojinzi/component/service/IServiceLifecycle.java
@@ -1,0 +1,6 @@
+package com.xiaojinzi.component.service;
+
+import com.xiaojinzi.component.support.IBaseLifecycle;
+
+public interface IServiceLifecycle extends IBaseLifecycle {
+}