+++ ComponentImpl/src/main/java/com/xiaojinzi/component/support/IBaseLifecycle.java
@@ -4,20 +4,19 @@ import android.app.Application;
 import android.support.annotation.MainThread;
 import android.support.annotation.NonNull;
 
-import com.xiaojinzi.component.ComponentConstants;
 import com.xiaojinzi.component.anno.support.CheckClassNameAnno;
 
 /**
- * 模块的生命周期接口
+ * 基础的生命周期接口
  */
 @MainThread
 @CheckClassNameAnno
-public interface IModuleLifecycle {
+public interface IBaseLifecycle {
 
     @MainThread
-    void onModuleCreate(@NonNull Application app);
+    void onCreate(@NonNull Application app);
 
     @MainThread
-    void onModuleDestroy();
+    void onDestroy();
 
 }