+++ Module2/src/main/java/com/xiaojinzi/component2/Component2Application.kt
@@ -2,10 +2,10 @@ package com.xiaojinzi.component2
 
 import android.app.Application
 import com.xiaojinzi.component.anno.ModuleAppAnno
-import com.xiaojinzi.component.application.IComponentApplication
+import com.xiaojinzi.component.application.IApplicationLifecycle
 
 @ModuleAppAnno()
-class Component2Application : IComponentApplication {
+class Component2Application : IApplicationLifecycle {
 
     override fun onCreate(app: Application) {
     }